+++ assets/js/src/public/product/variants.js
@@ -70,7 +70,6 @@ const setButtonState = (button) => {
 	disableActionButton(button);
 };
 
-
 /**
  * @function setInventory
  * @description Updates inventory/out of stock message and inputs
@@ -130,6 +129,56 @@ const handleAlertMessage = (formWrapper = '') => {
 	}
 
 	container.insertAdjacentHTML('beforeend', productMessage(state.variantMessage));
+};
+
+/**
+ * @function handleOptionStates
+ * @description Toggle the state of the current form options based on chosen options.
+ * @param formWrapper - Parent form node.
+ */
+const handleOptionStates = (formWrapper) => {
+	const options = tools.getNodes('product-form-option', false, formWrapper);
+
+	// Don't disable the first option so the user can choose alternative options.
+	if (options.length < 2) {
+		return;
+	}
+
+	options.forEach((option, i) => {
+		if (i === 0) return; // don't touch first option
+
+		const input = tools.getNodes('bc-product-option-field', true, option);
+
+		if (input && input[0]) {
+			let target = input[0];
+
+			if (input[0].type === 'checkbox' || input[0].type === 'radio') {
+				input.forEach((inputEl) => {
+					if (inputEl.checked) {
+						target = inputEl;
+					} else {
+						inputEl.removeAttribute('disabled');
+					}
+				});
+			} else if (input[0].type === 'select-one') {
+				tools.getNodes('option', true, input[0], true).forEach((inputEl) => {
+					if (inputEl.selected) {
+						target = inputEl;
+					} else {
+						inputEl.removeAttribute('disabled');
+					}
+				});
+			}
+
+			if (state.isValidOption) {
+				target.removeAttribute('disabled');
+			} else {
+				target.setAttribute('disabled', true);
+			}
+		}
+
+		option.classList.toggle('bc-product-form__option-variants--disabled');
+	});
 };
 
 /**
@@ -483,6 +532,7 @@ const handleSelections = (e, node = '') => {
 	setInventory();
 	showHideVariantImage(null, metaWrapper);
 	setButtonState(submitButton);
+	handleOptionStates(formWrapper);
 	handleAlertMessage(formWrapper);
 };