+++ templates/public/components/option-types/option-rectangles.php
@@ -17,7 +17,7 @@
 	<span class="bc-form__label bc-product-form__option-label <?php if ( $required ) { echo esc_attr( 'bc-form-control-required' ); } ?>"><?php echo esc_html( $label ); ?></span>
 
 	<!-- data-js="product-form-option" and data-field="product-form-option-radio" are required -->
-	<div class="bc-product-form__option-variants--inline" data-js="product-form-option" data-field="product-form-option-radio">
+	<div class="bc-product-form__option-variants bc-product-form__option-variants--inline" data-js="product-form-option" data-field="product-form-option-radio">
 		<?php foreach ( $options as $key => $option ) { ?>
 			<input type="radio"
 				name="option[<?php echo esc_attr( $id ); ?>]"